requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
babel==2.7.0              # via -r requirements.in, django-oscar, django-phonenumber-field
boto3==1.10.34            # via -r requirements.in
botocore==1.13.34         # via boto3, s3transfer
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
dj-database-url==0.5.0    # via -r requirements.in
django-appconf==1.0.3     # via -r requirements.in, django-compressor
django-compressor==2.3    # via -r requirements.in
django-extra-views==0.11.0  # via -r requirements.in, django-oscar
django-haystack==2.8.1    # via -r requirements.in, django-oscar
django-localflavor==2.2   # via django-oscar-paypal
django-nose==1.4.6        # via -r requirements.in
django-oscar-paypal==2.0.0  # via -r requirements.in
django-oscar==2.0.4       # via -r requirements.in
django-phonenumber-field==2.0.1  # via -r requirements.in, django-oscar
django-storages==1.8      # via -r requirements.in
django-tables2==1.21.2    # via -r requirements.in, django-oscar
django-treebeard==4.3     # via -r requirements.in, django-oscar
django-widget-tweaks==1.4.5  # via -r requirements.in, django-oscar
django==2.2.10            # via -r requirements.in, django-appconf, django-extra-views, django-haystack, django-localflavor, django-oscar, django-oscar-paypal, django-phonenumber-field, django-storages, django-tables2, django-treebeard
docutils==0.15.2          # via botocore
factory-boy==2.12.0       # via -r requirements.in, django-oscar
faker==3.0.0              # via -r requirements.in, factory-boy
gunicorn==20.0.4          # via -r requirements.in
idna==2.8                 # via requests
jmespath==0.9.4           # via boto3, botocore
mock==3.0.5               # via -r requirements.in
nose==1.3.7               # via django-nose
pbr==5.4.4                # via -r requirements.in
phonenumbers==8.11.0      # via -r requirements.in, django-oscar, django-phonenumber-field
phonenumberslite==8.11.0  # via -r requirements.in
pillow==6.2.1             # via -r requirements.in, django-oscar
pip-tools==4.3.0          # via -r requirements.in
psycopg2-binary==2.8.4    # via -r requirements.in
purl==1.5                 # via -r requirements.in, django-oscar
pycountry==19.8.18        # via -r requirements.in
python-dateutil==2.8.0    # via -r requirements.in, botocore, faker
python-stdnum==1.12       # via django-localflavor
pytz==2019.3              # via -r requirements.in, babel, django
rcssmin==1.0.6            # via -r requirements.in, django-compressor
requests==2.22.0          # via django-oscar-paypal
rjsmin==1.1.0             # via -r requirements.in, django-compressor
s3transfer==0.2.1         # via boto3
six==1.13.0               # via -r requirements.in, django-appconf, django-extra-views, faker, mock, pip-tools, purl, python-dateutil
sorl-thumbnail==12.5.0    # via -r requirements.in
sqlparse==0.3.0           # via django
text-unidecode==1.3       # via -r requirements.in, faker
unidecode==1.1.1          # via -r requirements.in
urllib3==1.25.7           # via botocore, requests
whitenoise==4.1.4         # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
